+++ CoreAlgorithmPlugins/src/au/gov/asd/tac/constellation/plugins/algorithms/sna/centrality/PathScoringUtilities.java
@@ -20,9 +20,10 @@ import au.gov.asd.tac.constellation.graph.GraphReadMethods;
 import au.gov.asd.tac.constellation.graph.schema.visual.concept.VisualConcept;
 import au.gov.asd.tac.constellation.utilities.datastructure.ThreeTuple;
 import au.gov.asd.tac.constellation.utilities.datastructure.Tuple;
-import java.util.ArrayList;
 import java.util.BitSet;
+import org.eclipse.collections.api.list.primitive.MutableFloatList;
 import org.eclipse.collections.api.list.primitive.MutableIntList;
+import org.eclipse.collections.impl.list.mutable.primitive.FloatArrayList;
 import org.eclipse.collections.impl.list.mutable.primitive.IntArrayList;
 
 /**
@@ -238,7 +239,7 @@ public class PathScoringUtilities {
     protected static ThreeTuple<BitSet[], float[], BitSet> computeAllPathsUndirectedThreeTuple(final GraphReadMethods graph, final ScoreType scoreType) {
         final int vxCount = graph.getVertexCount();
 
-        final ArrayList<Float> distances = new ArrayList<>();
+        final MutableFloatList distances = new FloatArrayList();
 
         final BitSet update = new BitSet(vxCount);
         final BitSet updateToReturn = new BitSet(vxCount);
@@ -343,7 +344,7 @@ public class PathScoringUtilities {
 
         final int vertexCount = graph.getVertexCount();
 
-        final ArrayList<Float> distances = new ArrayList<>();
+        final MutableFloatList distances = new FloatArrayList();
 
         final BitSet update = new BitSet(vertexCount);
         final BitSet updateReturn = new BitSet(vertexCount);
@@ -788,7 +789,7 @@ public class PathScoringUtilities {
         }
     }
 
-    private static void updateAveragePathScoresUndirected(final ArrayList<Float> distances, final float[] scores, final BitSet turn, final BitSet[] sendBuffer, final MutableIntList updateVertexArray, final GraphReadMethods graph) {
+    private static void updateAveragePathScoresUndirected(final MutableFloatList distances, final float[] scores, final BitSet turn, final BitSet[] sendBuffer, final MutableIntList updateVertexArray, final GraphReadMethods graph) {
         // for each node that has a message in transit, update its eccentricity
         for (int vxPos = turn.nextSetBit(0); vxPos >= 0; vxPos = turn.nextSetBit(vxPos + 1)) {
             int vxId = graph.getVertex(vxPos);