+++ CoreWordCloudView/src/au/gov/asd/tac/constellation/views/wordcloud/content/DefaultTokenHandler.java
@@ -15,8 +15,8 @@
  */
 package au.gov.asd.tac.constellation.views.wordcloud.content;
 
-import java.util.HashMap;
-import java.util.Map;
+import org.eclipse.collections.api.map.primitive.MutableObjectIntMap;
+import org.eclipse.collections.impl.map.mutable.primitive.ObjectIntHashMap;
 
 /**
  *
@@ -26,7 +26,7 @@ public class DefaultTokenHandler implements TokenHandler {
 
     @SuppressWarnings("unchecked") // type of SparseMatrix will be integer
     private final SparseMatrix<Integer> tokenElementMatrix = (SparseMatrix<Integer>) SparseMatrix.constructMatrix(0);
-    private final Map<String, Integer> tokenHashes = new HashMap<>();
+    private final MutableObjectIntMap<String> tokenHashes = new ObjectIntHashMap<>();
 
     public SparseMatrix<Integer> getTokenElementMatrix() {
         return tokenElementMatrix;